python/requirements.txt
adrf==0.1.8
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
attrs==23.2.0
blinker==1.8.2
brotli==1.1.0
certifi==2024.2.2
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
configargparse==1.7
cryptography==43.0.1
django==5.1.3
django-appconf==1.0.6
django-cors-headers==4.3.1
django-cryptography==1.1
django-environ==0.11.2
djangorestframework==3.15.1
drf-yasg==1.21.7
flask==3.0.3
flask-cors==5.0.0
flask-login==0.6.3
frozenlist==1.4.1
gevent==24.10.3
geventhttpclient==2.3.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.4
locust==2.32.0
markupsafe==3.0.2
msgpack==1.1.0
multidict==6.0.5
mysqlclient==2.2.4
onesignal-sdk==2.0.0
packaging==24.1
pillow==11.0.0
psutil==6.1.0
pycparser==2.22
pyjwt==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
pywin32==308; platform_system=='Windows'
pyyaml==6.0.2
pyzmq==26.2.0
razorpay==1.4.2
requests==2.32.2
setuptools==74.1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
twilio==9.1.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
werkzeug==3.0.6
yarl==1.9.4
zope-event==5.0
zope-interface==7.1.1
